VOIP-1470-Insight-AI-conversation-listen
- bin-ai-manager: Re-arm the conversation resolver with an EXPIRE-only touch instead of SADD so a concurrently stopped membership is never resurrected, cap the media tokens in a listen line, meter debounce lock errors, and split the intake tests into their own file
- docs: Reword the resolver re-arm implementation note
