+++ docs/plans/2026-09-05-insight-ai-conversation-listen-design.md
@@ -22,7 +22,7 @@ All `file:line` references were read against `main @ c243f926b` on 2026-09-05.
 | 4 | 2026-09-05 | Review round 3 (REQUEST_CHANGES, none blocking): `kind="unknown"` value for the pre-branch metric sites (M-1), §9 cells reconciled for the resolver quartet, `ListenPendingLen` and step 5b (M-2), §4 diagram and §5.3.2 step 5 folded into the per-AIcall loop (L-3), flush `ran` read against turn `skipped_empty` (L-4), §11 row 1 marked superseded (L-5), raw `media.Type` with all nine values (L-6), `publisherConversationManager` constant (L-7). Matrix in §11.2. |
 | 5 | 2026-09-05 | Review round 4 (APPROVE, 4 LOW): three more `kind` sites classified (`listen_trigger.go:254`, `listen.go:272` -> `unknown`; `listen_trigger.go:217` -> `call`) (L-1); §4 diagram corrected (pending is not trimmed) and a §6 row for the closed-Case, outgoing-only accumulation path with its bounds (L-2); §5.3.2 wrapper returns `error`, handler has no return (L-3); §11 row 10 annotated (L-4). Matrix in §11.3. |
 | 6 | 2026-09-05 | Review round 5 (APPROVE, 5 LOW, no re-review required): §5.1 call signature matched to §5.1.1 (L-1); intake-path `kind` attribution incl. `listen.go:720` and the `none -> unknown` rule (L-2, L-3); `failed` bucket for the step-4 `h.Get` error (L-4); `processEvent*` signature citation (L-5). Design APPROVED. Matrix in §11.4. |
-| 7 | 2026-09-05 | Implementation notes (VOIP-1470 PR): (a) the trigger's conversation branch additionally fetches the conversation via ConversationV1ConversationGet and refuses a cross-customer or nil-customer conversation with skipped_not_listenable (RPC error -> failed), a defence-in-depth check the intake tenant assertion already implied (code review finding); (b) listenKind helpers live in pkg/aicallhandler/listen_kind.go to keep listen.go under the 800-line guideline; (c) the terminate gate is the pure helper listenTerminateNeedsStop (ListenCallID OR listen pointer), unit-tested directly; (d) RunListenTurn's predicate-failed skipped_invalid emits the AIcall's actual kind (unknown only when no pointer exists); (e) intake meters stale resolver entries as dropped_stale (a resolved AIcall already over, or a pointer naming another conversation), keeping dropped_unknown for an unresolved or errored lookup; (f) the intake re-arms the conversation resolver membership and its 12h TTL on every buffered line, because start is the only SADD and a session can outlive listenResolverTTL (code review round 3). |
+| 7 | 2026-09-05 | Implementation notes (VOIP-1470 PR): (a) the trigger's conversation branch additionally fetches the conversation via ConversationV1ConversationGet and refuses a cross-customer or nil-customer conversation with skipped_not_listenable (RPC error -> failed), a defence-in-depth check the intake tenant assertion already implied (code review finding); (b) listenKind helpers live in pkg/aicallhandler/listen_kind.go to keep listen.go under the 800-line guideline; (c) the terminate gate is the pure helper listenTerminateNeedsStop (ListenCallID OR listen pointer), unit-tested directly; (d) RunListenTurn's predicate-failed skipped_invalid emits the AIcall's actual kind (unknown only when no pointer exists); (e) intake meters stale resolver entries as dropped_stale (a resolved AIcall already over, or a pointer naming another conversation), keeping dropped_unknown for an unresolved or errored lookup; (f) the intake re-arms the conversation resolver TTL (EXPIRE only, never SADD, so a concurrently stopped membership is never resurrected) on every buffered line, because start is the only SADD and a session can outlive listenResolverTTL (code review round 3, round 4). |
 
 ---
